requirements.txt
typer
openai
rich
python-dotenv
